+++ pychron/core/ui/label_editor.py
@@ -15,46 +15,11 @@
 # ===============================================================================
 
 # ============= enthought library imports =======================
-from pyface.tasks.action.task_action import TaskAction
 # ============= standard library imports ========================
 #============= local library imports  ==========================
-from pychron.envisage.resources import icon
+from pychron.core.ui.factory import toolkit_factory
 
-
-class AddNoteAction(TaskAction):
-    name = 'Add Note'
-    method = 'add_note'
-    image = icon('note-add')
-
-
-class SaveNoteAction(TaskAction):
-    name = 'Save Note'
-    method = 'save_note'
-    image = icon('document-save')
-
-
-class AddFolderAction(TaskAction):
-    name = 'Add Folder'
-    method = 'add_folder'
-    image = icon('folder-new')
-
-
-class PushAction(TaskAction):
-    name = 'Push'
-    method = 'push'
-    image = icon('arrow_up')
-
-
-class PullAction(TaskAction):
-    name = 'Pull'
-    method = 'pull'
-    image = icon('arrow_down')
-
-
-class NewLabelAction(TaskAction):
-    name = 'New Label'
-    method = 'new_label'
-    image = icon('add')
+LabelEditor = toolkit_factory('label_editor', 'LabelEditor')
 
 #============= EOF =============================================