+++ pychron/labbook/tasks/panes.py
@@ -16,23 +16,66 @@
 
 # ============= enthought library imports =======================
 from pyface.tasks.traits_dock_pane import TraitsDockPane
-from traits.api import HasTraits, Button, Str, Int, Bool, Any, List
 from traitsui.api import View, Item, UItem, HGroup, VGroup, TreeEditor, TreeNode
 # ============= standard library imports ========================
 # ============= local library imports  ==========================
-from traitsui.editors import ListStrEditor, TabularEditor, InstanceEditor
+from traitsui.editors import TabularEditor, InstanceEditor
 from traitsui.group import VSplit
 from traitsui.item import spring
+from traitsui.tabular_adapter import TabularAdapter
 from pychron.core.hierarchy import Hierarchy, FilePath, FilePathAdapter
 from pychron.envisage.icon_button_editor import icon_button_editor
 from pychron.git_archive.history import CommitAdapter
 
 
-class FileHistoryPane(TraitsDockPane):
-    name='File History'
-    id = 'pychron.labbook.file_history'
+class LabelTabularAdapter(TabularAdapter):
+    columns = [('Label', 'text')]
+
+    def get_bg_color(self, object, trait, row, column=0):
+        c = 'white'
+        cstr = self.item.color
+        if self.item.active:
+            c = '#{}'.format(cstr[:-2])
+
+        # print 'get bg color', self.item.text, self.item.active, c
+        return c
+
+    def get_text_color(self, obj, trait, row, column=0):
+
+        item = getattr(obj, trait)[row]
+
+        if item.active:
+            r = item.color[:2]
+            g = item.color[2:4]
+            b = item.color[4:6]
+            s = sum((int(r, 16), int(g, 16), int(b, 16)))
+            c = 'white' if s < 256 else 'black'
+        else:
+            c = '#{}'.format(item.color[:-2])
+
+        return c
+
+
+class LabelPane(TraitsDockPane):
+    name = 'Labels'
+    id = 'pychron.labbook.labels'
+
     def traits_view(self):
-        v=View(VGroup(
+        v = View(UItem('labels', editor=TabularEditor(editable=False,
+                                                      selected='selected',
+                                                      multi_select=True,
+                                                      dclicked='dclick',
+                                                      refresh='refresh_needed',
+                                                      adapter=LabelTabularAdapter())))
+        return v
+
+
+class FileHistoryPane(TraitsDockPane):
+    name = 'File History'
+    id = 'pychron.labbook.file_history'
+
+    def traits_view(self):
+        v = View(VGroup(
             HGroup(spring, Item('limit')),
             VSplit(UItem('items',
                          height=0.75,
@@ -66,7 +109,7 @@ class NotesBrowserPane(TraitsDockPane):
         v = View(VGroup(
             HGroup(Item('chronology_visible',
                         label='View Chronology'),
-                   icon_button_editor('filter_by_date_button','calendar'),
+                   icon_button_editor('filter_by_date_button', 'calendar'),
                    UItem('date_filter')),
             HGroup(Item('filter_hierarchy_str',
                         label='Name Filter')),